+++ ootrando_overworldmap_hamsda/scripts/logic_entrance.lua
@@ -454,3 +454,34 @@ function can_spawn_rainbow_bridge()
 
   return AccessibilityLevel.None
 end
+
+function can_trigger_lacs()
+  local setting_lacs_amount = get_object("setting_lacs_amount") and get_object("setting_lacs_amount").AcquiredCount or 0
+  local stones = Tracker:ProviderCountForCode("stones")
+  local medallions = Tracker:ProviderCountForCode("medallions")
+  local tokens = Tracker:ProviderCountForCode("token")
+
+  if has("setting_lacs_vanilla") then
+    if has("lacs_meds", 2) then
+      return AccessibilityLevel.Normal
+    end
+  elseif has("setting_lacs_stones") then
+    if stones >= setting_lacs_amount then
+      return AccessibilityLevel.Normal
+    end
+  elseif has("setting_lacs_medallions") then
+    if medallions >= setting_lacs_amount then
+      return AccessibilityLevel.Normal
+    end
+  elseif has("setting_lacs_dungeons") then
+    if (stones + medallions) >= setting_lacs_amount then
+      return AccessibilityLevel.Normal
+    end
+  elseif has("setting_lacs_gs") then
+    if tokens >= setting_lacs_amount then
+      return AccessibilityLevel.Normal
+    end
+  end
+
+  return AccessibilityLevel.None
+end